requirements.txt
streamlit
torch
torchvision
numpy
Pillow
opencv-python-headless
pytorch-grad-cam
matplotlib
scipy